+++ src/main/java/com/roloduck/user/model/UserRoleMapper.java
@@ -17,8 +17,8 @@ public class UserRoleMapper implements RowMapper<UserRole> {
     @Override
     public UserRole mapRow(ResultSet resultSet, int i) throws SQLException {
         UserRole userRole = new UserRole();
-        userRole.setId(resultSet.getInt("id"));
-        userRole.setUserId(resultSet.getInt("user_id"));
+        userRole.setId(resultSet.getLong("id"));
+        userRole.setUserId(resultSet.getLong("user_id"));
         userRole.setAuthority(resultSet.getString("authority"));
         return userRole;
     }
